### 2018-03-04 #600 @cMadan
Hi @plooney!

I took a quick look at your project repo, but it seems like the Readme isn't 'minimally complete' yet and the paper.md is still quite brief. While papers at JOSS are brief, they still need to sufficiently describe the project's purpose and the function it fulfills. If you haven't already, you should take a look at some recently accepted papers/projects as examples of the state they are in to be publishable at JOSS.

Let us know if you have any questions about the process!
