### 2018-03-06 1c85158
Adding a figure for the paper
