### 2018-03-06 #600 @plooney
I've added to the readme and the paper. What constitutes 'minimally complete'? I've not described the CMake steps in detail. A necessary requirement to use this is to have ITK installed so familiarity with the CMake build process is fair to assume.
