### 2018-03-06 #600 @cMadan
HI @plooney,

If you take a look at any of the review issues (e.g., https://github.com/openjournals/joss-reviews/issues/597), there is a list of what a reviewer will be asked to check for as part of the JOSS review. As your repo is now, many of these are currently missing or incomplete. E.g., no community guidelines, there is a test with an example input but not output, there is only a single sentence to address functionality, statement of need.

If you look at the repo for this submission, https://github.com/FastNFT/FNFT, you can see that the readme is significantly more detailed (admittedly you may not need OS-dependent installation instructions, but you can see the difference in the level of detail between the two).
