+++ README.md
@@ -13,7 +13,7 @@
 
 ** How to build **
 
-Use CMake and make
+Use cmake and make
 
 ** How to run tests **
 
@@ -22,3 +22,19 @@ make test
 ** How to create docs **
 
 make doc
+
+## Community Guidelines
+
+Please use the [issue tracker](https://github.com/plooney/kretz/issues) to report any problems with the software. If you want to contribute to the development of KRETZ, please send a pull request.
+
+## Contributors
+
+* Padraig Looney, University of Oxford
+
+## License
+
+KRETZ is provided under the terms of the [GNU Lesser General Public License, version 3](https://www.gnu.org/licenses/lgpl-3.0.en.html). 
+
+## Acknowledgements
+
+* This project has received funding from was supported by the Eunice Kennedy Shriver National Institute of Child Health and Human Development (NICHD) Human Placenta Project of the National Institutes of Health under award number UO1-HD087209.
