+++ README.md
@@ -4,6 +4,8 @@
 
 * KRETZ is a small library to enable reading a 3D Ultrasound Kretzfile and to convert between the toroidal geometry and cartesian space and visa versa 
 * Version 0.01
+* Two executabels are KretzWriter and KretzConverter to perform the conversion
+* The first three tests found in the CMakeList.txt file descibe the use and command line options of the executables to convert between the different geometries
 
 ### How do I get set up? ###
 
