### 2018-03-06 #600 @plooney
Hi @cMadab,

I've added to the ReadMe usinf the FNFT as a sample to address these issues.

I'm not clear on the point you make on the test. There are four tests listed in the CMakeLists with various input and output. "make test" will run all four and produce output and the input is also present. A sample image is downloaded and placed into the test folder that is used in the first three tests.
