### 2018-03-28 #600 @cMadan
For the tests, I think the output file should also be provided, so that the user can compare their output file with the pre-made output file to confirm that the code ran successfully.
