### 2018-04-12 a12f18f
Adding voluson output for comparison
