### 2018-06-18 #600 @lassoan
> This software was not inspired by your KretzFileReader

The two solutions are so similar, it's fascinating that we ended up implementing all this so similarly.

>  I will look to implementing them when time permits

Finding a suitable repository name and making sure the library can be built on all platforms would be important. Let me know what are the items in the mandatory list that you may not have time to implement and I'll try to help.

What is your plan in term of maintenance, support, developments in the future? Do you plan to review and accept pull request with improvements?
