### 2018-07-03 b2ae917
Adding doppler resampling
