### 2018-07-04 #600 @thewtex
Sorry I will not be able to follow up on this more until after I return from a conference the week after next.

Thanks @lassoan for the review and suggestions. Regarding

> File hosting on google drive is fragile. It is already broken (files are not available at the link provided) and even if you fix it, it is likely to break again. I would recommend to use the same test data hosting mechanism as all other ITK remote modules.

We currently host ITK module testing data on [data.kitware.com](https://data.kitware.com). @plooney  if you sign up for an account there, you can host files in your personal folder. Even better, let me know your account name, and I will add upload permissions to the [ITK Remote Module testing data folder](https://data.kitware.com/#collection/57b5c9e58d777f126827f5a1/folder/5817681d8d777f10f26ee354). This can be used [as a content link](https://itk.org/ITKExamples/Documentation/Contribute/UploadBinaryData.html) if an [External/Remote ITK module is created](https://github.com/InsightSoftwareConsortium/ITKModuleTemplate)

@stefanklein indicated he could possibly use this contribution to help with his research -- he may also be a good reviewer.
