+++ CMakeLists.txt
@@ -5,6 +5,7 @@ include(ExternalProject)
 project(ToroidalTransform)
 set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -Wall")
 set( CMAKE_EXPORT_COMPILE_COMMANDS ON )
+set(CMAKE_CXX_STANDARD 14)
 
 enable_testing()
 include(Dart)
