### 2018-09-17 #600 @cMadan
@lassoan has provided some detailed comments and suggestions, I would appreciate it if you could respond to each of these. @thewtex has also provided some suggestions for hosting test data, as well as provided the suggestion for a reviewer.
