### 2018-11-02 plooney/kretz#1 @thewtex
Use standard license file
The `LICENSE` file references LGPL 3, but using the standard LGPL 3 license file:

  https://choosealicense.com/licenses/lgpl-3.0/

will help [GitHub detect](https://help.github.com/articles/licensing-a-repository/), identify, and display the associated license.
