### 2018-11-02 #1063 @thewtex
>  License: Does the repository contain a plain-text LICENSE file with the contents of an OSI approved software license?

https://github.com/plooney/kretz/issues/1 created to clarify / standardize the license file.
