### 2018-11-02 plooney/kretz#2 @thewtex
Create a release
Create a Git tag for a stable version.

Describe the content of the release in the project [GitHub Release](https://github.com/plooney/kretz/releases).

Optionally add binaries for the release.
