### 2018-11-02 #1063 @thewtex
> Version: Does the release version given match the GitHub release (0.01)?

https://github.com/plooney/kretz/issues/2 created to create a GitHub release. It may be desirable to address other review items before creating this release.
