### 2018-11-02 plooney/kretz#1 @thewtex
Also, removing the *COPYRIGHT* and *COPYRIGHT.LESSER* will reduce ambiguity.
