### 2018-11-02 plooney/kretz#4 @thewtex
Missing `doc/kretz.doxygen` file
In CMake configuration, received

```
CMake Error: File /home/matt/src/kretz/doc/kretz.doxygen does not exist.
CMake Error at CMakeLists.txt:88 (configure_file):
  configure_file Problem configuring file

-- Configuring incomplete, errors occurred!
```
