### 2018-11-02 plooney/kretz#5 @thewtex
Override warnings with Clang 3.8.1 / ITK 4.13.1
The follow override warnings result with Clang 3.8.1 / ITK 4.13.1, indicating possible issues:

```
[3/8] Building CXX object CMakeFiles/ToroidalTransformTest.dir/test/itkToroidalTransformTest.cxx.o
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:120:25: warning: 'TransformPoint' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    OutputPointType     TransformPoint(const InputPointType  &point ) const;
                        ^
/home/matt/src/kretz/test/itkToroidalTransformTest.cxx:125:5: note: in instantiation of template class 'itk::ToroidalToCartesianTransform<double, 3>' requested here
    T2CTransformType::Pointer t2c = T2CTransformType::New();
    ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:189:27: note: overridden virtual function is here
  virtual OutputPointType TransformPoint(const InputPointType  &) const = 0;
                          ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:124:30: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVectorType TransformVector(const InputVectorType &) const
                             ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:192:29: note: overridden virtual function is here
  virtual OutputVectorType  TransformVector(const InputVectorType &) const
                            ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:132:33: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                                ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:207:31: note: overridden virtual function is here
  virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                              ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:140:39: warning: 'TransformCovariantVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputCovariantVectorType TransformCovariantVector(
                                      ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:238:37: note: overridden virtual function is here
  virtual OutputCovariantVectorType TransformCovariantVector(const InputCovariantVectorType &) const
                                    ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:149:18: warning: 'ComputeJacobianWithRespectToPosition' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void ComputeJacobianWithRespectToPosition( const InputPointType &, JacobianType &)  const
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:535:16: note: overridden virtual function is here
  virtual void ComputeJacobianWithRespectToPosition(const InputPointType & itkNotUsed(x), JacobianType & itkNotUsed(jacobian) )const
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:156:18: warning: 'ComputeJacobianWithRespectToParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void ComputeJacobianWithRespectToParameters( const InputPointType &, JacobianType &)  const
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:521:16: note: overridden virtual function is here
  virtual void ComputeJacobianWithRespectToParameters(const InputPointType  & itkNotUsed(p), JacobianType & itkNotUsed(jacobian) ) const = 0;
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:164:18: warning: 'SetParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void SetParameters(const ParametersType &){
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:346:16: note: overridden virtual function is here
  virtual void SetParameters(const ParametersType &) ITK_OVERRIDE = 0;
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:167:18: warning: 'SetFixedParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void SetFixedParameters(const FixedParametersType &){
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:381:16: note: overridden virtual function is here
  virtual void SetFixedParameters(const FixedParametersType &) ITK_OVERRIDE = 0;
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:3:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:176:10: warning: 'PrintSelf' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    void PrintSelf(std::ostream &os, Indent indent) const;
         ^
/home/matt/src/ITK/Modules/Core/Common/include/itkObject.h:192:16: note: overridden virtual function is here
  virtual void PrintSelf(std::ostream & os, Indent indent) const ITK_OVERRIDE;
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:118:25: warning: 'TransformPoint' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    OutputPointType     TransformPoint(const InputPointType  &point ) const;
                        ^
/home/matt/src/kretz/test/itkToroidalTransformTest.cxx:126:5: note: in instantiation of template class 'itk::CartesianToToroidalTransform<double, 3>' requested here
    C2TTransformType::Pointer c2t = C2TTransformType::New();
    ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:189:27: note: overridden virtual function is here
  virtual OutputPointType TransformPoint(const InputPointType  &) const = 0;
                          ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:122:30: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVectorType TransformVector(const InputVectorType &) const
                             ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:192:29: note: overridden virtual function is here
  virtual OutputVectorType  TransformVector(const InputVectorType &) const
                            ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:130:33: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                                ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:207:31: note: overridden virtual function is here
  virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                              ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:138:39: warning: 'TransformCovariantVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputCovariantVectorType TransformCovariantVector(
                                      ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:238:37: note: overridden virtual function is here
  virtual OutputCovariantVectorType TransformCovariantVector(const InputCovariantVectorType &) const
                                    ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:147:18: warning: 'ComputeJacobianWithRespectToPosition' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void ComputeJacobianWithRespectToPosition( const InputPointType &, JacobianType &)  const
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:535:16: note: overridden virtual function is here
  virtual void ComputeJacobianWithRespectToPosition(const InputPointType & itkNotUsed(x), JacobianType & itkNotUsed(jacobian) )const
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:155:18: warning: 'ComputeJacobianWithRespectToParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void ComputeJacobianWithRespectToParameters( const InputPointType &, JacobianType &)  const
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:521:16: note: overridden virtual function is here
  virtual void ComputeJacobianWithRespectToParameters(const InputPointType  & itkNotUsed(p), JacobianType & itkNotUsed(jacobian) ) const = 0;
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:164:18: warning: 'SetParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void SetParameters(const ParametersType &){
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:346:16: note: overridden virtual function is here
  virtual void SetParameters(const ParametersType &) ITK_OVERRIDE = 0;
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:167:18: warning: 'SetFixedParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void SetFixedParameters(const FixedParametersType &){
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:381:16: note: overridden virtual function is here
  virtual void SetFixedParameters(const FixedParametersType &) ITK_OVERRIDE = 0;
               ^
In file included from /home/matt/src/kretz/test/itkToroidalTransformTest.cxx:4:
/home/matt/src/kretz/include/itkCartesianToToroidalTransform.h:176:10: warning: 'PrintSelf' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    void PrintSelf(std::ostream &os, Indent indent) const;
         ^
/home/matt/src/ITK/Modules/Core/Common/include/itkObject.h:192:16: note: overridden virtual function is here
  virtual void PrintSelf(std::ostream & os, Indent indent) const ITK_OVERRIDE;
               ^
18 warnings generated.
[5/8] Building CXX object CMakeFiles/KretzFileWriter.dir/src/itkKretzWriter.cxx.o
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:120:25: warning: 'TransformPoint' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    OutputPointType     TransformPoint(const InputPointType  &point ) const;
                        ^
/home/matt/src/kretz/src/itkKretzWriter.cxx:52:5: note: in instantiation of template class 'itk::ToroidalToCartesianTransform<double, 3>' requested here
    T2CTransformType::Pointer t2c = T2CTransformType::New();
    ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:189:27: note: overridden virtual function is here
  virtual OutputPointType TransformPoint(const InputPointType  &) const = 0;
                          ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:124:30: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVectorType TransformVector(const InputVectorType &) const
                             ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:192:29: note: overridden virtual function is here
  virtual OutputVectorType  TransformVector(const InputVectorType &) const
                            ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:132:33: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                                ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:207:31: note: overridden virtual function is here
  virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                              ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:140:39: warning: 'TransformCovariantVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputCovariantVectorType TransformCovariantVector(
                                      ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:238:37: note: overridden virtual function is here
  virtual OutputCovariantVectorType TransformCovariantVector(const InputCovariantVectorType &) const
                                    ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:149:18: warning: 'ComputeJacobianWithRespectToPosition' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void ComputeJacobianWithRespectToPosition( const InputPointType &, JacobianType &)  const
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:535:16: note: overridden virtual function is here
  virtual void ComputeJacobianWithRespectToPosition(const InputPointType & itkNotUsed(x), JacobianType & itkNotUsed(jacobian) )const
               ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:156:18: warning: 'ComputeJacobianWithRespectToParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void ComputeJacobianWithRespectToParameters( const InputPointType &, JacobianType &)  const
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:521:16: note: overridden virtual function is here
  virtual void ComputeJacobianWithRespectToParameters(const InputPointType  & itkNotUsed(p), JacobianType & itkNotUsed(jacobian) ) const = 0;
               ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:164:18: warning: 'SetParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void SetParameters(const ParametersType &){
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:346:16: note: overridden virtual function is here
  virtual void SetParameters(const ParametersType &) ITK_OVERRIDE = 0;
               ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:167:18: warning: 'SetFixedParameters' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual void SetFixedParameters(const FixedParametersType &){
                 ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:381:16: note: overridden virtual function is here
  virtual void SetFixedParameters(const FixedParametersType &) ITK_OVERRIDE = 0;
               ^
In file included from /home/matt/src/kretz/src/itkKretzWriter.cxx:6:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:176:10: warning: 'PrintSelf' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    void PrintSelf(std::ostream &os, Indent indent) const;
         ^
/home/matt/src/ITK/Modules/Core/Common/include/itkObject.h:192:16: note: overridden virtual function is here
  virtual void PrintSelf(std::ostream & os, Indent indent) const ITK_OVERRIDE;
               ^
9 warnings generated.
[7/8] Building CXX object CMakeFiles/KretzConverter.dir/src/itkKretzConverter.cxx.o
In file included from /home/matt/src/kretz/src/itkKretzConverter.cxx:14:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:120:25: warning: 'TransformPoint' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    OutputPointType     TransformPoint(const InputPointType  &point ) const;
                        ^
/home/matt/src/kretz/src/itkKretzConverter.cxx:71:74: note: in instantiation of template class 'itk::ToroidalToCartesianTransform<double, 3>' requested here
BoundingBoxType::BoundsArrayType computeBounds(ImageType::Pointer image, T2CTransformType::Pointer t2c)
                                                                         ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:189:27: note: overridden virtual function is here
  virtual OutputPointType TransformPoint(const InputPointType  &) const = 0;
                          ^
In file included from /home/matt/src/kretz/src/itkKretzConverter.cxx:14:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:124:30: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVectorType TransformVector(const InputVectorType &) const
                             ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:192:29: note: overridden virtual function is here
  virtual OutputVectorType  TransformVector(const InputVectorType &) const
                            ^
In file included from /home/matt/src/kretz/src/itkKretzConverter.cxx:14:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:132:33: warning: 'TransformVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                                ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:207:31: note: overridden virtual function is here
  virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
                              ^
In file included from /home/matt/src/kretz/src/itkKretzConverter.cxx:14:
/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:140:39: warning: 'TransformCovariantVector' overrides a member function but is not marked 'override' [-Winconsistent-missing-override]
    virtual OutputCovariantVectorType TransformCovariantVector(
                                      ^
/home/matt/src/ITK/Modules/Core/Transform/include/itkTransform.h:238:37: note: overridden virtual function is here
[…]
