### 2018-11-02 #1063 @thewtex
>  Installation: Does installation proceed as outlined in the documentation?

Generally, yes, but a few minor issues identified:

- https://github.com/plooney/kretz/issues/4
- https://github.com/plooney/kretz/issues/5
