### 2018-11-02 plooney/kretz#6 @thewtex
Describe command line program options
Please provide a detailed description and example usage of command line program options.

It is currently a little unclear what the options do and how they should be specified. Please provide a more detailed description for the command line flags. Also, additional tests that exercise and demonstrate the flags would be helpful.
