### 2018-11-02 plooney/kretz#7 @thewtex
Document the difference between *KretzConverter* and *KretzFileWriter*
Please describe the difference between these two programs and the use case where one should be used over the other in the *README*.
