### 2018-11-02 plooney/kretz#8 @thewtex
Format as an ITK module to enable use as a library
In order for other projects to use the code, it should be made available as an ITK Module. This will provide the required CMake configuration information and ITK target dependency information. This will allow other projects to build against the C++ classes provided.

For more information on generation of the required files, see the [ITKModuleTemplate](https://github.com/InsightSoftwareConsortium/ITKModuleTemplate) and documentation on [How To Create A Module](https://itk.org/ITKSoftwareGuide/html/Book1/ITKSoftwareGuide-Book1ch9.html#x55-1490009) in the ITK Software Guide.
