### 2018-11-02 #1063 @thewtex
> Functionality: Have the functional claims of the software been confirmed?

https://github.com/plooney/kretz/issues/8 created to enable use as a library.
