### 2018-11-02 #1063 @thewtex
>  Functionality documentation: Is the core functionality of the software documented to a satisfactory level (e.g., API method documentation)?

There is Doxygen documentation and the naming of classes and methods are information.

A few items to improve with the command line programs:

- https://github.com/plooney/kretz/issues/6
- https://github.com/plooney/kretz/issues/7
