### 2018-11-02 plooney/kretz#10 @thewtex
Correct SlicerHeart reference
Correct `[@SklicerHeart] ` spelling in *paper.md* and provide a link to the project.

CC: @lassoan
