### 2018-11-02 #1063 @thewtex
>  References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?

- Minor correction required: https://github.com/plooney/kretz/issues/10
