### 2018-11-02 #1063 @thewtex
@cMadan @plooney review complete.

I apologize again for the tardiness of the review.

@plooney the code looks outstanding, and this is a valuable contribution for ultrasound research. Thank you for sharing this work.
