### 2018-11-19 4aed61c
Removing reference to IJMacros.txt
