### 2018-11-19 948b1d6
Adding to CMakeLists.txt
