### 2018-11-20 6f07a2c
Fixing gh-#5 issue with CLang  and adding doc/kretz.doxygen gh-#4
