+++ CMakeLists.txt
@@ -2,9 +2,12 @@ cmake_minimum_required(VERSION 2.8)
 include(ExternalProject)
 
 project(ToroidalTransform)
-set(CMAKE_CXX_FLAGS_DEBUG "${CMAKE_CXX_FLAGS_DEBUG} -Wall")
+set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall")
 set( CMAKE_EXPORT_COMPILE_COMMANDS ON )
 set(CMAKE_CXX_STANDARD 14)
+if(USE_CLANG)
+	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fpermissive -Wno-inconsistent-missing-override -Wno-missing-braces")
+endif(USE_CLANG)
 
 enable_testing()
 include(Dart)
