### 2018-11-21 1432907
Adding info on command line option gh-#6 and difference between executables gh-#7
