+++ README.md
@@ -19,21 +19,21 @@ Use cmake and make
 
 **Programs**
 
-KretzFileWriter either takes a KRETZ file and outputs the voxels in the geometry of the 3D ultrasound probe. One dimension corresponds to radial distance and the others correspond to the angles. THe spacing is not isotropic and can be found in the KRETZ file. KretzFileWriter can take a cartesian geometry and a KRETZ file and convert the cartesian geometry back into the geometry specified in the KRETZ file.
+KretzFileWriter either takes a KRETZ file and outputs the voxels in the geometry of the 3D ultrasound probe. One dimension corresponds to radial distance and the others correspond to the angles. The spacing is not isotropic and can be found in the KRETZ file. KretzFileWriter can take a cartesian geometry and a KRETZ file and convert the cartesian geometry back into the geometry specified in the KRETZ file.
 
-arguments: 
-    i - input KRETZ file
-    c - optional cartesian file to convert back into geometry specified in the KRETZ file
-    o - path for the output file
+  *arguments: 
+   - i - input KRETZ file
+   - c - optional cartesian file to convert back into geometry specified in the KRETZ file 
+   - o - path for the output file
 
 KretzConverter takes a KRETZ file and converts it to cartesian coordinates.
-    i - input KRETZ file
-    o - path for the output file
-    r - three floating point values corresponding to the voxel spacing in each direction, if not specified s must be
-    s - three integers to define the number of voxels in each direction, if not specifed r must be 
-    m - create a mask image 1 where the voxel is in the geometry of the ultrasound beam and 0 otherwise
-    n - normalise voxel values to have zero mean and unit variance
-    d - write out power doppler instead of grayscale voxel values to geometry of the grayscale acquisition
+   - i - input KRETZ file
+   - o - path for the output file
+   - r - three floating point values corresponding to the voxel spacing in each direction, if not specified s must be
+   - s - three integers to define the number of voxels in each direction, if not specifed r must be 
+   - m - create a mask image 1 where the voxel is in the geometry of the ultrasound beam and 0 otherwise
+   - n - normalise voxel values to have zero mean and unit variance
+   - d - write out power doppler instead of grayscale voxel values to geometry of the grayscale acquisition
 
     
     
