### 2018-11-28 e3d12b2
Trying to allow building as an ITK External Module - Outstanding issue with error - undefined reference to itk::KretzImageIOFactoryRegister__Private() gh-#8
