+++ CMakeLists.txt
@@ -1,7 +1,9 @@
-cmake_minimum_required(VERSION 2.8)
-include(ExternalProject)
+cmake_minimum_required(VERSION 3.5.1 FATAL_ERROR)
+cmake_policy(VERSION 3.5.1)
 
-project(ToroidalTransform)
+project(ITK3DUltrasound)
+
+set(ToroidalTransform_LIBRARIES ToroidalTransform)
 set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -Wall")
 set( CMAKE_EXPORT_COMPILE_COMMANDS ON )
 set(CMAKE_CXX_STANDARD 14)
@@ -9,81 +11,97 @@ if(USE_CLANG)
 	set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -fpermissive -Wno-inconsistent-missing-override -Wno-missing-braces")
 endif(USE_CLANG)
 
-enable_testing()
-include(Dart)
+if(NOT ITK_SOURCE_DIR)
+	find_package(ITK REQUIRED)
+	list(APPEND CMAKE_MODULE_PATH ${ITK_CMAKE_DIR})
+	#include(ITKModuleExternal)
 
-find_package( Boost 1.58 COMPONENTS program_options REQUIRED )
-find_package(ITK REQUIRED)
-include_directories(src include ${Boost_INCLUDE_DIR} )
+	enable_testing()
+	include(Dart)
 
-set(Libraries ${ITK_LIBRARIES})
+	find_package( Boost 1.58 COMPONENTS program_options REQUIRED )
+	find_package(ITK REQUIRED)
+	include_directories(src include ${Boost_INCLUDE_DIR} )
 
-set(SOURCE
-	src/itkKretzWriter.cxx
-	src/itkKretzImageIO.cxx
-	)
-include( ${ITK_USE_FILE} )
-add_executable(KretzFileWriter ${SOURCE} )
-target_link_libraries(KretzFileWriter ${} ${ITK_LIBRARIES} ${Boost_LIBRARIES})
+	set(Libraries ${ITK_LIBRARIES})
 
-add_executable(KretzConverter
-	src/itkKretzConverter.cxx
-	src/itkKretzImageIO.cxx
-	src/itkCartesianToToroidalTransform.txx
-	src/itkToroidalToCartesianTransform.txx
-	)
+	set(SOURCE
+		src/itkKretzWriter.cxx
+		src/itkKretzImageIO.cxx
+		#src/itkCartesianToToroidalTransform.cxx
+		#src/itkToroidalToCartesianTransform.cxx
+		)
+	include( ${ITK_USE_FILE} )
 
-target_link_libraries(KretzConverter ${Libraries} ${Boost_LIBRARIES})
+	add_executable(KretzFileWriter ${SOURCE} )
+	target_link_libraries(KretzFileWriter ${ITK_LIBRARIES} ${Boost_LIBRARIES})
 
-if(BUILD_TESTING)
-	add_executable(ToroidalTransformTest
-		src/itkToroidalToCartesianTransform.txx
-		src/itkCartesianToToroidalTransform.txx
-		test/itkToroidalTransformTest.cxx
+	add_executable(KretzConverter
+		src/itkKretzConverter.cxx
+		src/itkKretzImageIO.cxx
+		#src/itkCartesianToToroidalTransform.cxx
+		#src/itkToroidalToCartesianTransform.cxx
 		)
 
-	target_link_libraries(ToroidalTransformTest ${Libraries})
+	target_link_libraries(KretzConverter ${Libraries} ${Boost_LIBRARIES})
 
-	# Test writing the toroidal volume out to a nii.gz file format, geometry of the file is determined by angles in  
[…]
