+++ src/CMakeLists.txt
@@ -1,0 +1,6 @@
+set(3DUltrasound_SRCS
+	itkKretzImageIO.cxx 
+	itkKretzImageIOFactory.cxx
+	)
+
+itk_module_add_library(ITK3DUltrasound ${3DUltrasound_SRCS})
