### 2018-11-28 093535a
Adding travis CI file
