+++ .travis.yml
@@ -15,7 +15,7 @@ cache:
 matrix:
   include:
     - os: linux
-      dist: trusty
+      dist: xenial 
       sudo: required
 
 before_script:
