+++ .travis.yml
@@ -27,7 +27,7 @@ before_script:
 
 script:
   - cd ${KRETZ_BUILD_DIR}
-  - cmake -DWRAP_LUA:BOOL=OFF -DWRAP_PYTHON:BOOL=OFF -DWRAP_JAVA:BOOL=OFF -DWRAP_CSHARP:BOOL=OFF -DWRAP_TCL:BOOL=OFF -DWRAP_R:BOOL=OFF -DWRAP_RUBY:BOOL=OFF -DBUILD_EXAMPLES:BOOL=OFF -DBUILD_TESTING:BOOL=OFF -DBUILD_DOXYGEN:BOOL=ON -DCMAKE_BUILD_TYPE=Release "${KRETZ_SRC_DIR}"
-  - make docs 
+  - cmake -DWRAP_LUA:BOOL=OFF -DWRAP_PYTHON:BOOL=OFF -DWRAP_JAVA:BOOL=OFF -DWRAP_CSHARP:BOOL=OFF -DWRAP_TCL:BOOL=OFF -DWRAP_R:BOOL=OFF -DWRAP_RUBY:BOOL=OFF -DBUILD_EXAMPLES:BOOL=OFF -DBUILD_TESTING:BOOL=OFF -DBUILD_DOXYGEN:BOOL=ON -DCMAKE_BUILD_TYPE=Release  -fpermissive -Wno-inconsistent-missing-override -Wno-missing-braces "${KRETZ_SRC_DIR}" 
+  - make doc 
   - make  
   - make test
