### 2018-12-05 12680e5
Adding ITK build to travis
