### 2018-12-10 8fcaa36
Correcting SlicerHeart ref gh-#10
