+++ paper.bib
@@ -1,6 +1,6 @@
 @online{SlicerHeart,
   author = {SlicerHeart},
-  title = {SlicerHeart extension for 3D Slicer},
+  title = {3D Slicer extension for cardiac analysis},
   year = 2017,
   url = {https://github.com/SlicerHeart/SlicerHeart},
   urldate = {2018-02-27}
