### 2018-12-10 3b86f18
Removing COPYRIGHT* files and moving to version 1.0 gh-#2
