### 2018-12-10 #1063 @plooney
@cMadan

Thanks for the comments @thewtex. I have tried to address all the points, I struggled to get the two of the issues resolved fully; ITK Module and CI.
