### 2018-12-15 plooney/kretz#2 @thewtex
@plooney at a minimum, a [Git tag](https://git-scm.com/book/en/v2/Git-Basics-Tagging) should be created to identify the release. A tag identifies a specific state of the code (an identifier that points to the SHA1 commit hash identifying the state of the repository) that can be referenced, which is important for reproducibility.
