### 2018-12-15 0f1598a
COMP: Add missing override's
To address:

/home/matt/src/kretz/include/itkToroidalToCartesianTransform.h:120:25:
warning: 'TransformPoint' overrides a member function but is not marked
'override' [-Winconsistent-missing-override]
    OutputPointType     TransformPoint(const InputPointType  &point )
    const;
