+++ include/itkToroidalToCartesianTransform.h
@@ -117,11 +117,11 @@ public:
    * This method transforms first two dimensions of a point from polar
    * coordinates <alpha,radius> to cartesian coordinates.
    */
-    OutputPointType     TransformPoint(const InputPointType  &point ) const;
+    OutputPointType TransformPoint(const InputPointType  &point ) const override;
 
     /** Method to transform a vector - not applicable for this type of
       transform. */
-    virtual OutputVectorType TransformVector(const InputVectorType &) const
+    OutputVectorType TransformVector(const InputVectorType &) const override
     {
         itkExceptionMacro(<< "Method not applicable for polar transform." );
         return OutputVectorType();
@@ -129,7 +129,7 @@ public:
 
     /** Method to transform a vnl_vector - not applicable for this type of
       transform. */
-    virtual OutputVnlVectorType TransformVector(const InputVnlVectorType &) const
+    OutputVnlVectorType TransformVector(const InputVnlVectorType &) const override
     {
         itkExceptionMacro(<< "Method not applicable for polar transform. ");
         return OutputVnlVectorType();
@@ -137,8 +137,8 @@ public:
 
     /** Method to transform a CovariantVector - not applicable for this type of
       transform */
-    virtual OutputCovariantVectorType TransformCovariantVector(
-            const InputCovariantVectorType &) const
+    OutputCovariantVectorType TransformCovariantVector(
+            const InputCovariantVectorType &) const override
     {
         itkExceptionMacro(<< "Method not applicable for polar transfrom. ");
         return OutputCovariantVectorType();
@@ -146,14 +146,14 @@ public:
 
     /** Compute the Jacobian Matrix of the transformation at one point - not
       applicable for this type of transform */
-    virtual void ComputeJacobianWithRespectToPosition( const InputPointType &, JacobianType &)  const
+    void ComputeJacobianWithRespectToPosition( const InputPointType &, JacobianType &)  const override
     {
         itkExceptionMacro(<< "Method not applicable for polar transform. ");
     }
 
     /** Compute the Jacobian Matrix of the transformation at one point - not
       applicable for this type of transform */
-    virtual void ComputeJacobianWithRespectToParameters( const InputPointType &, JacobianType &)  const
+    void ComputeJacobianWithRespectToParameters( const InputPointType &, JacobianType &)  const override
     {
         itkExceptionMacro(<< "Method not applicable for polar transform. ");
     }
@@ -161,10 +161,10 @@ public:
     virtual void SetTableTheta(const TableType);
     virtual void SetTablePhi(const TableType);
 
-    virtual void SetParameters(const ParametersType &){
+    void SetParameters(const ParametersType &) override {
 
     }
-    virtual void SetFixedParameters(const FixedParametersType &){
+    void SetFixedParameters(const FixedParametersType &) override {
 
     }
 
@@ -173,7 +173,7 @@ protected:
     ~ToroidalToCartesianTransform();
 
     /** Print contents of an ToroidalToCartesianTransform. */
-    void PrintSelf(std::ostream &os, Indent indent) const;
+    void PrintSelf(std::ostream &os, Indent indent) const override;
 
 private:
     ToroidalToCartesianTransform(const Self&); //purposely not implemented
