### 2018-12-15 f442d9b
ENH: Add infrastructure to build as an ITK module
This enables CI testing as use as a library.

This contains the output of the ITK module template:

  https://github.com/InsightSoftwareConsortium/ITKModuleTemplate

Running:

  python -m pip install cookiecutter
  python -m cookiecutter gh:InsightSoftwareConsortium/ITKModuleTemplate

By conventions for ITK IO modules, the name is "IOKretz". The CMake
project name and libraries correspond to this name.

The example applications are moved in the "examples" directory per ITK
module convention.

Fix style issues identified by IOKretzKWStyleTest.


[…]
