### 2018-12-15 25b6c3d
ENH: Add CI testing for the programs
