+++ .circleci/config.yml
@@ -36,6 +36,15 @@ jobs:
           no_output_timeout: 1.0h
           command: |
             ctest -j 2 -VV -S dashboard.cmake
+      - run:
+          name: Test example programs
+          no_output_timeout: 1.0h
+          command: |
+            mkdir /examples-build
+            cd /examples-build
+            cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DITK_DIR=/ITK-build /kretz/examples/
+            ninja
+            ctest
 
 workflows:
     version: 2
