### 2018-12-15 #1063 @thewtex
@plooney looking good.

Please review and fix up this PR as required:

  https://github.com/plooney/kretz/pulls

which should how to format as an ITK Module to use as a library and adds CI configuration.

One other item is to tag the repository.
