### 2018-12-21 plooney/kretz#8 @plooney
@thewtex If i try to build outside the ITK buildpath I get

CMake Error at CMakeLists.txt:22 (include):
   include could not find load file:

     ITKModuleExternal
