### 2018-12-22 plooney/kretz#8 @thewtex
Did you try these steps?

```
mkdir kretz-module-build
cd kretz-module-build
cmake -DITK_DIR=/path/to/ITK-build /path/to/kretz
make
cd ..

mkdir kretz-programs-build
cd kretz-programs-build
cmake -DITK_DIR=/path/to/ITK-build /path/to/kretz/examples
make
```
