### 2019-01-08 e1fa93e
adding version of c++ to examples CMAkeLists.txt
