+++ examples/CMakeLists.txt
@@ -1,5 +1,6 @@
 cmake_minimum_required(VERSION 3.10.2 FATAL_ERROR)
 cmake_policy(VERSION 3.10.2)
+set(CMAKE_CXX_STANDARD 14)
 
 project(ITK3DUltrasound)
 
