+++ .circleci/config.yml
@@ -35,6 +35,7 @@ jobs:
           name: Build and Test
           no_output_timeout: 1.0h
           command: |
+            cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DITK_DIR=/ITK-build -DBUILD_TESTING:BOOL=ON
             ctest -j 2 -VV -S dashboard.cmake
       - run:
           name: Test example programs
