+++ .circleci/config.yml
@@ -35,7 +35,9 @@ jobs:
           name: Build and Test
           no_output_timeout: 1.0h
           command: |
-            ctest -j 2 -VV -S dashboard.cmake
+            cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DITK_DIR=/ITK-build -DBUILD_TESTING:BOOL=ON /kretz
+            ninja
+            ctest
       - run:
           name: Test example programs
           no_output_timeout: 1.0h
