+++ .circleci/config.yml
@@ -8,6 +8,8 @@ jobs:
       - checkout:
           path: /kretz
       - run:
+          command: 'sudo apt-get install libboost-dev'
+      - run:
           name: Fetch CTest driver script
           command: |
             curl -L https://raw.githubusercontent.com/InsightSoftwareConsortium/ITK/dashboard/itk_common.cmake -O
@@ -43,7 +45,6 @@ jobs:
           no_output_timeout: 1.0h
           command: |
             mkdir /examples-build
-            sudo apt-get install libboost-dev
             cd /examples-build
             cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DITK_DIR=/ITK-build /kretz/examples/
             ninja
