+++ .circleci/config.yml
@@ -8,6 +8,7 @@ jobs:
       - checkout:
           path: /kretz
       - run:
+          name: install
           command: 'sudo apt-get install libboost-dev'
       - run:
           name: Fetch CTest driver script
