+++ .circleci/config.yml
@@ -9,7 +9,7 @@ jobs:
           path: /kretz
       - run:
           name: install
-          command: 'aptitude install libboost-dev'
+          command: 'rpm install libboost-dev'
       - run:
           name: Fetch CTest driver script
           command: |
