### 2019-01-09 9111bbe
Update config.yml
Need to find a way to add boost to test the examples. Commented out for now
