+++ .circleci/config.yml
@@ -42,15 +42,15 @@ jobs:
             cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DITK_DIR=/ITK-build -DBUILD_TESTING:BOOL=ON /kretz
             ninja
             ctest
-      - run:
-          name: Test example programs
-          no_output_timeout: 1.0h
-          command: |
-            mkdir /examples-build
-            cd /examples-build
-            cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DITK_DIR=/ITK-build /kretz/examples/
-            ninja
-            ctest
+#      - run:
+#          name: Test example programs
+#          no_output_timeout: 1.0h
+#          command: |
+#            mkdir /examples-build
+#            cd /examples-build
+#            cmake -G Ninja -DCMAKE_BUILD_TYPE=Release -DITK_DIR=/ITK-build /kretz/examples/
+#            ninja
+#            ctest
 
 workflows:
     version: 2
