### 2019-01-11 #1063 @plooney
I changed the tag to v1.1 to include the latest commit with all the responses to the review.
